Добавит правки наставника
